+++ src/inspect_ai/_util/dict.py
@@ -1,0 +1,9 @@
+from typing import Any
+
+
+def omit(x: dict[str, Any], vars: list[str]) -> dict[str, Any]:
+    x = x.copy()
+    for var in vars:
+        if var in x:
+            del x[var]
+    return x