+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Set Claude default `max_tokens` to 4096
 - Combine user and assistant messages for Vertex models.
 - Warn when using the `name` parameter with task created from `@task` decorated function.
+- Make sample `metadata` available in prompt, grading, and self-criqique templates.
 
 ## v0.3.20 (03 August 2024)